FIX: Feature space (#58)
* fix 1 selected ft - no transformation performed2

* postprocess data transform in mlflow eval

* fix abundance_threshold min and max

* fix nn_reg type error with boolean md enrichment

+++ ritme/evaluate_models.py
@@ -201,8 +201,7 @@ class TunedModel:
         other_ft_ls, enriched = self.enrich(
             data, micro_ft_raw, transformed, self.data_config, split
         )
-
-        X = enriched[micro_ft_transf + other_ft_ls]
+        X = enriched[micro_ft_transf + other_ft_ls].astype(float)
 
         if isinstance(self.model, NeuralNet):
             self.model.eval()
